poc-application-code/threetierapp/middletierapp/requirements.txt
flask
flask_restful
pandas
mysql-connector-python